Resolved the -Wsign-conversion warnings in QIR tests

+++ src/Qir/Runtime/lib/Simulators/ToffoliSimulator.cpp
@@ -35,7 +35,7 @@ namespace Quantum
 
         static uint64_t GetQubitId(Qubit qubit)
         {
-            return static_cast<uint64_t>((uintptr_t)qubit);
+            return (uint64_t)qubit;
         }
 
       public:
